Expand all groups
Re-save the file with all groups expanded, to minimize the occurrence of
unexpected messages indicating that the database has been changed.
Apparently the folder state (collapsed or expanded) is saved as part of
the database, and expanding a folder will cause KeePassX to consider the
item to have been edited. This is very counterintuitive, to say the very
least.